Increase version to 0.1.4
This commit includes a fix for a UnicodeDecodeError exception being raised when
building the pbldr package in a clean chroot. The error was occurring when
trying to read in the long_description from README.rst. The exception was only
occuring in the clean chroot environment, and I have no clue why. This patch
does fix it.